requirements.txt
pylint
Pillow
